Integrate files
